upy-fs-builder: Replicate MicroPython behaviour with extra 0xFF byte
When the last chunk is used up to the last byte the micro:bit
MicroPython file system will also use the next chunk and leave it
empty (the double link list pointers and end file offset will still
be calculated like there is an extra 0xFF at the end of data).
This commit replicates that behaviour by adding an extra 0xFF byte
at the end of the data.