upy-fs-builder: Add generateHexWithFiles() for quicker Hex file generation.
This version uses a pre-cached struct of data with the MicroPython
Intel hex string and Memory map to be able to generate a full Intel
Hex string with MicroPython and filesystem much quicker.

On a MacBook with CPU throtling inthe chrome debugger we can save
about 3s the hex generation, which is how long it took to constantly
convert the MicroPython MemoryMap data into Intel Hex records.